Move CompletableResultCode to datadog.trace.api
